requirements.txt
django==3.1.4
djangorestframework==3.12.2
django-cors-headers==3.5.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
django-rest-passwordreset==1.1.0
